Zero Acquaintance Round Robin Data From Albirght, Kenny, and Malloy

Data from Study 3 of Albright et al. (1988)

data(Zero)

Format

A data frame with 636 measurements of 36 round robin groups on the following 15 round-robin variables (taken on 7-point rating scales with higher values indicating more of the trait):

sociable: rating of dimension sociable-reclusive
good: rating of dimension good-natured-irritable
responsible: rating of dimension responsible-undependable
calm: rating of dimension calm-anxious
intellectual: rating of dimension intellectual-unintellectual
imaginative: rating of dimension imaginative-unimaginative
talkative: rating of dimension talkative-silent
fussy: rating of dimension fussy-careless
composed: rating of dimension composed-excitable
cooperative: rating of dimension cooperative-negativistic
physically_attractive: rating of dimension physically attractive-unattractive
formal_dress: rating of dimension formal dress-casual dress
neatly_dressed: rating of dimension neatly dressed-sloppy dress
athletic: rating of dimension athletic-not athletic
young: rating of dimension young-old

The data frame also contains the gender (actor.sex; 1 = F, 2 = M) of the participants and their self-ratings on the five assessed traits (actor.sociable and so on).

Source

http://davidakenny.net/srm/srmdata.htm

References

Albright, L., Kenny, D. A., & Malloy, T. E. (1988). Consensus in personality judgments at zero acquaintance. Journal of Personality and Social Psychology, 55(3), 387-395. doi: 10.1037/0022-3514.55.3.387