Structural Equation Modeling for the Social Relations Model

Provides functionality for structural equation modeling for the social relations model (Kenny & La Voie, 1984; <doi:10.1016/S0065-2601(08)60144-6>; Warner, Kenny, & Soto, 1979, <doi:10.1037/0022-3514.37.10.1742>). Maximum likelihood estimation (Gill & Swartz, 2001, <doi:10.2307/3316080>; Nestler, 2018, <doi:10.3102/1076998617741106>) and least squares estimation is supported (Bond & Malloy, 2018, <doi:10.1016/B978-0-12-811967-9.00014-X>).

References

Bond, C. F., & Malloy, T. E. (2018a). Social relations analysis of dyadic data structure: The general case. In T. E. Malloy. Social relations modeling of behavior in dyads and groups (Ch. 14). Academic Press. doi: 10.1016/B978-0-12-811967-9.00014-X

Gill, P. S., & Swartz, T. B. (2001). Statistical analyses for round robin interaction data. Canadian Journal of Statistics, 29(2), 321-331. doi: 10.2307/3316080

Kenny, D. A., & La Voie, L. J. (1984). The social relations model. In L. Berkowitz (Ed.), Advances in experimental social psychology (Vol. 18, pp. 142-182). Orlando, FL: Academic. doi: 10.1016/S0065-2601(08)60144-6

Nestler, S. (2018). Likelihood estimation of the multivariate social relations model. Journal of Educational and Behavioral Statistics, 43(4), 387-406. doi: 10.3102/1076998617741106

Warner, R. M., Kenny, D. A., & Soto, M. (1979). A new round robin analysis of variance for social interaction data. Journal of Personality and Social Psychology, 37(10), 1742-1757. doi: 10.1037/0022-3514.37.10.1742

See also