Hallmark and Kenny Round Robin Data

Data from Kenny et al. (1994)

data(HallmarkKenny)

Format

A data frame with 802 measurements of 30 round-robin groups on the following 7 round-robin variables (taken on unnumbered 7-point rating scales with higher numbers indicating a higher value of the trait):

calm: rating of dimension calm-anxious
sociable rating of dimension sociable-withdrawn
liking rating of dimension like-do not like
careful rating of dimension careful-careless
relaxed rating of dimension relaxed-tense
talkative rating of dimension talkative-quiet
responsible rating of dimension responsible-undependable

The data frame also contains participants gender (actor.sex; 1 = F, 2 = M) and their age in years (actor.age). Note that the data was assessed in two conditions: odd round robin group numbers indicate groups in which participants rated all traits for a person at a time whereas even numbers refer to groups in which participants rated all the people for each trait.

Source

http://davidakenny.net/srm/srmdata.htm

References

Kenny, D. A., Albright, L., Malloy, T. E., & Kashy, D. A. (1994). Consensus in interpersonal perception: Acquaintance and the big five. Psychological Bulletin, 116(2), 245-258. doi: 10.1037/0033-2909.116.2.245